refactor: isolate review retry command formatting
