+++ packages/cli/src/review/coordinator.ts
@@ -1,6 +1,7 @@
 import { resolveRunIdentity } from '../../templates/hooks/lib/run-identity.js';
 import type { ProgressReporter } from '../cli-protocol/handler.js';
 import { type CliResult, createResult, type Effect, type Finding } from '../cli-protocol/result.js';
+import { retryCommand } from './command.js';
 import type {
   ReviewAgent,
   ReviewAuthor,
@@ -164,26 +165,6 @@ function assessFallback(
   return provenance.kind === 'failed'
     ? { kind: 'failed', failure: provenance.code, terminal: false }
     : { kind: 'completed', output: provenance.output };
-}
-
-function shellQuote(value: string): string {
-  if (/^[\w./-]+$/u.test(value)) return value;
-  const escaped = value.replaceAll("'", `'"'"'`);
-  return `'${escaped}'`;
-}
-
-export function retryCommand(
-  kind: ReviewKind,
-  targets: readonly string[],
-  context: readonly string[] = [],
-): string {
-  // `--` ends option parsing, so a reviewed file named `--help` or `-r` reaches
-  // the command as a target rather than as a flag. Shell quoting cannot do this:
-  // it is the argument parser, not the shell, that would misread the name.
-  const quoted = targets.map(target => shellQuote(target)).join(' ');
-  const contextOption =
-    context.length === 0 ? '' : ` --context ${context.map(target => shellQuote(target)).join(' ')}`;
-  return `safeword review run ${kind}${contextOption} -- ${quoted}`;
 }
 
 /** How an agent is written for a reader: the product name, not the runtime id. */
